Move the sync part to a separate package
