+++ internal/usersync/queries.go
@@ -3,8 +3,6 @@ package usersync
 import (
 	"context"
 
-	"github.com/google/uuid"
-	"github.com/nais/api/internal/graph/ident"
 	"github.com/nais/api/internal/graph/pagination"
 	"github.com/nais/api/internal/usersync/usersyncsql"
 )
@@ -26,16 +24,4 @@ func ListLogEntries(ctx context.Context, page *pagination.Pagination) (*UserSync
 	}
 
 	return pagination.NewConvertConnectionWithError(ret, page, total, toGraphUserSyncLogEntry)
-}
-
-func Get(ctx context.Context, uid uuid.UUID) (UserSyncLogEntry, error) {
-	return fromContext(ctx).userSyncLogLoader.Load(ctx, uid)
-}
-
-func GetByIdent(ctx context.Context, id ident.Ident) (UserSyncLogEntry, error) {
-	uid, err := parseIdent(id)
-	if err != nil {
-		return nil, err
-	}
-	return Get(ctx, uid)
 }
