+++ internal/usersync/usersyncer/usersyncer_test.go
@@ -1,6 +1,6 @@
 //go:build integration_test
 
-package usersync_test
+package usersyncer_test
 
 import (
 	"context"
@@ -15,7 +15,7 @@ import (
 	"github.com/nais/api/internal/role/rolesql"
 	"github.com/nais/api/internal/test"
 	"github.com/nais/api/internal/user"
-	"github.com/nais/api/internal/usersync"
+	"github.com/nais/api/internal/usersync/usersyncer"
 	"github.com/sirupsen/logrus"
 	logrustest "github.com/sirupsen/logrus/hooks/test"
 	"github.com/testcontainers/testcontainers-go/modules/postgres"
@@ -61,7 +61,7 @@ func TestSync(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		err = usersync.
+		err = usersyncer.
 			New(pool, adminGroupPrefix, domain, svc, log).
 			Sync(ctx)
 		if err != nil {
@@ -110,7 +110,7 @@ func TestSync(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		err = usersync.
+		err = usersyncer.
 			New(pool, adminGroupPrefix, domain, svc, log).
 			Sync(ctx)
 		if err != nil {
@@ -173,7 +173,7 @@ func TestSync(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		err = usersync.
+		err = usersyncer.
 			New(pool, adminGroupPrefix, domain, svc, log).
 			Sync(ctx)
 		if err != nil {
